ConfigApp: Split Misc tab into subtabs & rename to Settings
